+++ src/main/java/wooteco/subway/section/api/dto/SectionRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.line;
+package wooteco.subway.section.api.dto;
 
 public class SectionRequest {
